+++ lib/src/dialog_message.dart
@@ -1,61 +1,111 @@
 part of 'dialogs.dart';
 
+typedef CustomMessageDialogBuilder = Widget Function(
+  BuildContext context,
+  String? title,
+  String? message,
+);
+
 class MessageDialogConfig extends DialogConfig {
+  final CustomMessageDialogBuilder? builder;
+
   const MessageDialogConfig({
-    super.message,
+    super.material,
     super.messageStyle,
-    super.title,
     super.titleStyle,
     String? buttonText,
     TextStyle? buttonTextStyle,
-  }) : super(
+  })  : builder = null,
+        super(
           positiveButtonText: buttonText,
           positiveButtonTextStyle: buttonTextStyle,
         );
 
-  MessageDialogConfig copy({
-    String? buttonText,
-    TextStyle? buttonTextStyle,
-    String? message,
-    TextStyle? messageStyle,
-    String? title,
-    TextStyle? titleStyle,
-  }) {
-    return MessageDialogConfig(
-      buttonText: buttonText ?? positiveButtonText,
-      buttonTextStyle: buttonTextStyle ?? positiveButtonTextStyle,
-      message: message ?? this.message,
-      messageStyle: messageStyle ?? this.messageStyle,
-      title: title ?? this.title,
-      titleStyle: titleStyle ?? this.titleStyle,
-    );
-  }
+  const MessageDialogConfig.builder(
+    CustomMessageDialogBuilder this.builder, {
+    super.material,
+  }) : super();
 }
 
-class _MessageDialog extends StatefulWidget {
+class _MaterialMessageDialog extends StatelessWidget {
+  final String? title;
+  final String? message;
   final MessageDialogConfig config;
 
-  const _MessageDialog({
-    super.key,
+  const _MaterialMessageDialog({
+    required this.title,
+    required this.message,
     required this.config,
   });
 
   @override
-  State<_MessageDialog> createState() => _MessageDialogState();
+  Widget build(BuildContext context) {
+    if (config.builder != null) {
+      return config.builder!(context, title, message);
+    }
+    return Dialog(
+      clipBehavior: Clip.antiAlias,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(20),
+      ),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Container(
+            padding: const EdgeInsets.only(
+              top: 24,
+              bottom: 12,
+              left: 32,
+              right: 32,
+            ),
+            child: Text(
+              message ?? "",
+              textAlign: TextAlign.center,
+              overflow: TextOverflow.ellipsis,
+              maxLines: 5,
+            ),
+          ),
+          DialogButton(
+            material: true,
+            text: config.positiveButtonText ?? "OK",
+            style: config.positiveButtonTextStyle,
+            onClick: (_) => Navigator.pop(_, true),
+          ),
+        ],
+      ),
+    );
+  }
 }
 
-class _MessageDialogState extends State<_MessageDialog> {
+class _CupertinoMessageDialog extends StatefulWidget {
+  final String? title;
+  final String? message;
+  final MessageDialogConfig config;
+
+  const _CupertinoMessageDialog({
+    required this.title,
+    required this.message,
+    required this.config,
+  });
+
+  @override
+  State<_CupertinoMessageDialog> createState() =>
+      _CupertinoMessageDialogState();
+}
+
+class _CupertinoMessageDialogState extends State<_CupertinoMessageDialog> {
   late MessageDialogConfig config = widget.config;
 
   Widget? _title(BuildContext context) {
-    final data = config.title ?? "";
+    final data = widget.title ?? "";
     final isValid = data.isNotEmpty;
     if (!isValid) return null;
     return Text(data, style: config.titleStyle);
   }
 
   Widget? _subtitle(BuildContext context) {
-    final value = config.message ?? "";
+    final value = widget.message ?? "";
     final isValid = value.isNotEmpty;
     if (!isValid) return null;
     return Padding(
@@ -78,11 +128,11 @@ class _MessageDialogState extends State<_MessageDialog> {
 
   @override
   Widget build(BuildContext context) {
+    if (config.builder != null) {
+      return config.builder!(context, widget.title, widget.message);
+    }
     return BackdropFilter(
-      filter: ImageFilter.blur(
-        sigmaX: 10,
-        sigmaY: 10,
-      ),
+      filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
       child: CupertinoAlertDialog(
         actions: [_button(context)],
         content: _subtitle(context),