+++ lib/src/type.dart
@@ -3,14 +3,23 @@ part of 'dialogs.dart';
 enum DialogType {
   alert,
   editor,
-  loading,
-  message;
+  loader,
+  message,
+  snackBar,
+  snackBarError,
+  snackBarWarning;
 
   bool get isAlert => this == alert;
 
   bool get isEditor => this == editor;
 
-  bool get isLoading => this == loading;
+  bool get isLoader => this == loader;
 
   bool get isMessage => this == message;
+
+  bool get isSnackBar => this == snackBar;
+
+  bool get isErrorSnackBar => this == snackBarError;
+
+  bool get isWarningSnackBar => this == snackBarWarning;
 }