+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* Customizable all dialog
+* Initial custom widget support
+* Check multiple visibility
+* Add SnackBars (Error, Warning, Normal) like toast and use also custom
+
 ## 0.0.1
 
 * Add alert message dialog